Move galaxy_tasks decorator to galaxy.celery
That means we can import additional tasks that use `galaxy_task` in
galaxy.celery.tasks without circular imports.
